basic server for testing
