+++ config/test.json
@@ -1,0 +1,4 @@
+{
+  "port": 9090,
+  "db": "inspire-test"
+}
