+++ index.js
@@ -1,0 +1,3 @@
+import server from './bin/server'
+
+server()
