+++ package.json
@@ -3,9 +3,9 @@
   "version": "0.1.0",
   "description": "An API for the INSPIRE Index Polygons dataset. Look up an Inspire id for a given lng lat",
   "scripts": {
-    "start": "babel-node ./server.js",
-    "watch": "nodemon ./server.js --exec babel-node",
-    "test": "standard"
+    "start": "babel-node ./index.js",
+    "watch": "nodemon ./index.js --exec babel-node",
+    "test": "standard && NODE_ENV=test babel-node tests/*.js"
   },
   "repository": {
     "type": "git",
@@ -21,6 +21,7 @@
     "apollo-server": "^0.2.8",
     "body-parser": "^1.15.2",
     "boom": "^4.2.0",
+    "config": "^1.24.0",
     "express": "4.13.4",
     "graphql": "^0.7.2",
     "graphql-tools": "^0.7.2",
@@ -38,6 +39,7 @@
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "6.5.0",
     "nodemon": "^1.9.1",
-    "standard": "^8.6.0"
+    "standard": "^8.6.0",
+    "tape": "^4.6.3"
   }
 }
